+++ src/file_request_handler.cc
@@ -39,12 +39,12 @@
 
 #include "cds/cds_item.h"
 #include "config/config_manager.h"
+#include "config/result/transcoding.h"
 #include "content/content_manager.h"
 #include "database/database.h"
 #include "iohandler/file_io_handler.h"
 #include "metadata/metadata_handler.h"
 #include "transcoding/transcode_dispatcher.h"
-#include "transcoding/transcoding.h"
 #include "upnp_common.h"
 #include "upnp_xml.h"
 #include "util/grb_net.h"
