+++ test/core/test_upnp_xml.cc
@@ -24,10 +24,10 @@ Gerbera - https://gerbera.io/
 #include "cds/cds_container.h"
 #include "cds/cds_item.h"
 #include "common.h"
-#include "config/client_config.h"
+#include "config/result/client_config.h"
+#include "config/result/transcoding.h"
 #include "config/config_setup.h"
 #include "metadata/metadata_handler.h"
-#include "transcoding/transcoding.h"
 #include "upnp_xml.h"
 #include "util/grb_net.h"
 #include "util/tools.h"
